+++ docs/source/HISTORY.rst
@@ -4,6 +4,8 @@ Changelog
 0.2.12 (unreleased)
 -------------------
 - added progressbar for zip_upload()
+- added support for importing a single file through the 
+  ZIP import form into the current subdirectory^
 
 0.2.11 (2014-11-08)
 -------------------